renaming files to align with current conventions
