+++ packages/web-components/fast-ssr/src/styles/fast-style.ts
